create property groups file
