+++ bpy_speckle/connector/ui/project_selection_dialog.py
@@ -9,6 +9,7 @@ from ..utils.account_manager import (
     can_create_project_in_workspace,
 )
 from ..utils.project_manager import get_projects_for_account
+from ..utils.property_groups import speckle_project
 
 
 def get_accounts_callback(self, context):
@@ -33,18 +34,6 @@ def get_workspaces_callback(self, context):
         (workspace.id, workspace.name, "", "WORKSPACE", i)
         for i, workspace in enumerate(wm.speckle_workspaces)
     ]
-
-
-class speckle_project(bpy.types.PropertyGroup):
-    """
-    PropertyGroup for storing project information
-    """
-
-    name: bpy.props.StringProperty()  # type: ignore
-    role: bpy.props.StringProperty(name="Role")  # type: ignore
-    updated: bpy.props.StringProperty(name="Updated")  # type: ignore
-    id: bpy.props.StringProperty(name="ID")  # type: ignore
-    can_receive: bpy.props.BoolProperty(name="Can Receive", default=False)  # type: ignore
 
 
 class SPECKLE_UL_projects_list(bpy.types.UIList):
